Fix benchmark publishability warning scope
