+++ src/pythinker_code/benchmark/report.py
@@ -68,13 +68,6 @@ def render_run_report(
         estimated_cost = f"${float(usage['estimated_cost_usd']):.4f}"
     verification_output = _verification_output(verification)
     final_excerpt = _excerpt(final_answer)
-    from pythinker_code.benchmark.commands import BenchmarkReportRow
-    from pythinker_code.benchmark.compare import readiness_warnings
-
-    warnings = readiness_warnings(
-        [BenchmarkReportRow(run=dict(run), summary=dict(summary))]
-    )
-
     lines = [
         "# Pythinker Benchmark",
         "",
@@ -94,8 +87,6 @@ def render_run_report(
         f"- Artifacts: {artifact_root}",
         "",
     ]
-    if warnings:
-        lines.extend(["Publishability warnings:", *(f"- {warning}" for warning in warnings), ""])
     if has_valid_activity:
         lines.extend(["- Activity:", *activity_lines])
     else:
