+++ src/pythinker_code/benchmark/types.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+JsonObject = dict[str, object]
+
+
+@dataclass(frozen=True, slots=True)
+class BenchmarkReportRow:
+    run: JsonObject
+    summary: JsonObject
